Revert "Add "Download Map" Top Sheet card"
