--- OsmAnd/res/anim/slide_in_top.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<set xmlns:android="http://schemas.android.com/apk/res/android">
-	<translate android:fromYDelta="-100%p" android:toYDelta="0"
-		android:duration="@android:integer/config_mediumAnimTime"/>
-	<alpha android:fromAlpha="0.0" android:toAlpha="1.0"
-		android:duration="@android:integer/config_mediumAnimTime" />
-</set>
