--- OsmAnd/res/anim/slide_out_top.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<set xmlns:android="http://schemas.android.com/apk/res/android">
-	<translate android:fromYDelta="0" android:toYDelta="-100%p"
-		android:duration="@android:integer/config_mediumAnimTime"/>
-	<alpha android:fromAlpha="1.0" android:toAlpha="0.0"
-		android:duration="@android:integer/config_mediumAnimTime" />
-</set>
