--- OsmAnd/res/drawable/bg_top_sheet_bottom_sides_landscape_dark.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
-	<item>
-		<nine-patch android:src="@drawable/bg_shadow_dialogtop_landscape"/>
-	</item>
-	<item>
-		<shape>
-			<solid android:color="@color/list_background_color_dark"/>
-			<corners android:bottomRightRadius="3dp"
-				android:bottomLeftRadius="3dp" />
-		</shape>
-	</item>
-</layer-list>
