--- OsmAnd/res/layout/card_download_detailed_map.xml
@@ -1,72 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-	xmlns:tools="http://schemas.android.com/tools"
-	android:layout_width="match_parent"
-	android:layout_height="wrap_content"
-	xmlns:osmand="http://schemas.android.com/apk/res-auto"
-	android:background="?attr/list_background_color"
-	android:orientation="vertical">
-
-	<LinearLayout
-		android:layout_width="match_parent"
-		android:layout_height="wrap_content"
-		android:layout_marginLeft="@dimen/list_content_padding"
-		android:layout_marginRight="@dimen/list_content_padding"
-		android:orientation="vertical">
-
-		<LinearLayout
-			android:layout_width="match_parent"
-			android:layout_height="wrap_content"
-			android:orientation="horizontal">
-
-			<net.osmand.plus.widgets.TextViewEx
-				android:id="@+id/description"
-				android:layout_width="0dp"
-				android:layout_height="wrap_content"
-				osmand:typeface="@string/font_roboto_regular"
-				android:layout_marginTop="@dimen/list_content_padding"
-				android:layout_weight="1"
-				android:textSize="@dimen/default_list_text_size"
-				tools:text="@string/download_detaile_map" />
-
-			<ImageView
-				android:layout_width="wrap_content"
-				android:layout_height="wrap_content"
-				android:layout_marginLeft="@dimen/card_content_padding_large"
-				android:src="@drawable/img_download" />
-
-		</LinearLayout>
-
-		<net.osmand.plus.widgets.TextViewEx
-			android:id="@+id/fileSize"
-			android:layout_width="wrap_content"
-			android:layout_height="wrap_content"
-			osmand:typeface="@string/font_roboto_medium"
-			android:textColor="?attr/active_color_basic"
-			android:textSize="@dimen/default_desc_text_size"
-			tools:text="139mb" />
-
-		<LinearLayout
-			android:layout_width="match_parent"
-			android:layout_height="wrap_content"
-			android:layout_marginTop="@dimen/content_padding_small"
-			android:layout_marginBottom="@dimen/content_padding_small">
-
-			<include
-				android:id="@+id/btnClose"
-				layout="@layout/bottom_sheet_dialog_button" />
-
-			<View
-				android:id="@+id/buttons_divider"
-				android:layout_width="@dimen/content_padding"
-				android:layout_height="match_parent" />
-
-			<include
-				android:id="@+id/btnDownload"
-				layout="@layout/bottom_sheet_dialog_button" />
-
-		</LinearLayout>
-
-	</LinearLayout>
-
-</LinearLayout>
