+++ OsmAnd/res/values/styles.xml
@@ -647,11 +647,6 @@
         <item name="android:windowExitAnimation">@anim/shrink_from_top</item>
     </style>
 
-    <style name="Animations.PopUpMenu.Top">
-        <item name="android:windowEnterAnimation">@anim/slide_in_top</item>
-        <item name="android:windowExitAnimation">@anim/slide_out_top</item>
-    </style>
-    
     <style name="Animations.PopUpMenu.Bottom">
         <item name="android:windowEnterAnimation">@anim/slide_in_bottom</item>
         <item name="android:windowExitAnimation">@anim/slide_out_bottom</item>
