+++ OsmAnd/src/net/osmand/plus/AppInitializer.java
@@ -500,6 +500,7 @@ public class AppInitializer implements IProgress {
 		app.travelDbHelper = startupInit(app.travelDbHelper, TravelDbHelper.class);
 		app.lockHelper = startupInit(new LockHelper(app), LockHelper.class);
 
+
 		initOpeningHoursParser();
 	}
