+++ OsmAnd/src/net/osmand/plus/base/BottomSheetDialog.java
@@ -1,6 +1,5 @@
 package net.osmand.plus.base;
 
-import android.app.Activity;
 import android.app.Dialog;
 import android.content.Context;
 import android.content.res.TypedArray;
@@ -8,39 +7,31 @@ import android.os.Build;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
-import android.view.Gravity;
 import android.view.MotionEvent;
 import android.view.View;
-import android.view.ViewGroup;
 import android.view.ViewGroup.LayoutParams;
 import android.view.Window;
 import android.view.WindowManager;
 import android.widget.FrameLayout;
 
-import static net.osmand.plus.base.SheetDialogType.*;
+import net.osmand.plus.R;
 
-public class OsmandSheetDialog extends Dialog {
-	
-	public final static int CONTAINER_ID = 111111;
+public class BottomSheetDialog extends Dialog {
 
 	private boolean cancelable = true;
 	private boolean canceledOnTouchOutside = true;
-	private boolean interactWithOutside = false;
 	private boolean canceledOnTouchOutsideSet;
-	
-	private SheetDialogType dialogType;
 
-	public OsmandSheetDialog(@NonNull Context context) {
-		this(context, 0, BOTTOM);
+	public BottomSheetDialog(@NonNull Context context) {
+		this(context, 0);
 	}
 
-	public OsmandSheetDialog(@NonNull Context context, int themeResId, SheetDialogType dialogType) {
+	public BottomSheetDialog(@NonNull Context context, int themeResId) {
 		super(context, themeResId);
-		this.dialogType = dialogType;
 		requestWindowFeature(Window.FEATURE_NO_TITLE);
 	}
 
-	protected OsmandSheetDialog(@NonNull Context context, boolean cancelable, @Nullable OnCancelListener cancelListener) {
+	protected BottomSheetDialog(@NonNull Context context, boolean cancelable, @Nullable OnCancelListener cancelListener) {
 		super(context, cancelable, cancelListener);
 		requestWindowFeature(Window.FEATURE_NO_TITLE);
 		this.cancelable = cancelable;
@@ -89,23 +80,11 @@ public class OsmandSheetDialog extends Dialog {
 		canceledOnTouchOutside = cancel;
 		canceledOnTouchOutsideSet = true;
 	}
-	
-	public void setInteractWithOutside(boolean interactWithOutside) {
-		this.interactWithOutside = interactWithOutside;
-	}
 
 	@NonNull
 	private View wrapInContainer(int layoutResId, View view, LayoutParams params) {
-		final FrameLayout res = new FrameLayout(getContext());
-		final View touchOutside = new View(getContext());
-		final FrameLayout container = getContainer(dialogType);
-
-		ViewGroup.LayoutParams fullScreenParams = new ViewGroup.LayoutParams(LayoutParams.MATCH_PARENT, LayoutParams.MATCH_PARENT);
-		res.setLayoutParams(fullScreenParams);
-		touchOutside.setLayoutParams(fullScreenParams);
-		
-		res.addView(touchOutside);
-		res.addView(container);
+		final View res = View.inflate(getContext(), R.layout.bottom_sheet_dialog, null);
+		final FrameLayout container = (FrameLayout) res.findViewById(R.id.content_container);
 
 		if (layoutResId != 0 && view == null) {
 			view = getLayoutInflater().inflate(layoutResId, container, false);
@@ -116,25 +95,12 @@ public class OsmandSheetDialog extends Dialog {
 			container.addView(view, params);
 		}
 
-		touchOutside.setOnClickListener(new View.OnClickListener() {
+		res.findViewById(R.id.touch_outside).setOnClickListener(new View.OnClickListener() {
 			@Override
 			public void onClick(View view) {
 				if (cancelable && isShowing() && shouldWindowCloseOnTouchOutside()) {
 					cancel();
 				}
-			}
-		});
-		
-		touchOutside.setOnTouchListener(new View.OnTouchListener() {
-			@Override
-			public boolean onTouch(View v, MotionEvent event) {
-				if (interactWithOutside) {
-					Activity ownerActivity = getOwnerActivity();
-					if (ownerActivity != null) {
-						ownerActivity.dispatchTouchEvent(event);
-					}
-				}
-				return false;
 			}
 		});
 
@@ -147,29 +113,6 @@ public class OsmandSheetDialog extends Dialog {
 		});
 
 		return res;
-	}
-	
-	private FrameLayout getContainer(SheetDialogType dialogType) {
-		if (dialogType == null) {
-			return null;
-		}
-		int width = LayoutParams.MATCH_PARENT;
-		int height = LayoutParams.WRAP_CONTENT;
-		int gravity = 0;
-		switch (dialogType) {
-			case TOP:
-				gravity = Gravity.TOP | Gravity.CENTER_HORIZONTAL;
-				break;
-			case BOTTOM:
-				gravity = Gravity.BOTTOM | Gravity.CENTER_HORIZONTAL;
-				break;
-		}
-		FrameLayout.LayoutParams params = new FrameLayout.LayoutParams(width, height);
-		params.gravity = gravity;
-		FrameLayout container = new FrameLayout(getContext());
-		container.setLayoutParams(params);
-		container.setId(CONTAINER_ID);
-		return container;
 	}
 
 	private boolean shouldWindowCloseOnTouchOutside() {
