--- OsmAnd/src/net/osmand/plus/base/TopSheetDialogFragment.java
@@ -1,9 +1,0 @@
-package net.osmand.plus.base;
-
-public abstract class TopSheetDialogFragment extends OsmAndSheetDialogFragment {
-	
-	@Override
-	protected SheetDialogType getSheetDialogType() {
-		return SheetDialogType.TOP;
-	}
-}
